Fix and add unit tests
- Prefixed every fixture with "CkTest", because if you have a project with a users table and fixtures (very common) this package and any plugin having unit/integration tests that are executed after the tests of this package produced conflicts.
That's because the users table is not changed back to the fixture/schema of the main app after the tests of this package and therefor potentially missing some fields that are only present in the users table of the main app.
To easily prevent this, I addded the "CkTest" prefix to every fixture this package uses, minimalizing the potential of naming conflicts with real apps.
- Added unit tests for the latest configuration value of the StrictPasswordBehavior: "minUserNameLength", which sets a lower limit for the length of name fields of the user to be considered when checking that the password does not contain any of the names. This prevents people with very short names getting totally random passwords rejected because their short name appears in the random characters of their desired new password